+++ SCSRaytracer/src/Materials/PerfectSpecular.cs
@@ -19,29 +19,29 @@ namespace RayTracer
 {
     class PerfectSpecular : BRDF
     {
-        private double kr;
+        private float kr;
         private RGBColor cr;
 
         public PerfectSpecular()
         {
-            kr = 1.0;
-            cr = new RGBColor(1.0, 1.0, 1.0);
+            kr = 1.0f;
+            cr = new RGBColor(1.0f, 1.0f, 1.0f);
         }
-        public PerfectSpecular(RGBColor color, double kr_arg)
+        public PerfectSpecular(RGBColor color, float kr_arg)
         {
             kr = kr_arg;
             cr = color;
         }
 
-        public void setKr(double kr_arg) { kr = kr_arg; }
+        public void setKr(float kr_arg) { kr = kr_arg; }
         public void setCr(RGBColor cr_arg) { cr = cr_arg; }
-        public double getKr() { return kr; }
+        public float getKr() { return kr; }
         public RGBColor getCr() { return cr; }
 
         public override RGBColor sample_f(ShadeRec sr, ref Vect3D wi, ref Vect3D wo)
         {
-            double ndotwo = sr.normal * wo;
-            wi = -wo + 2.0 * sr.normal * ndotwo;
+            float ndotwo = sr.normal * wo;
+            wi = -wo + 2.0f * sr.normal * ndotwo;
 
             return (kr * cr / (sr.normal * wi));
         }
