+++ SCSRaytracer/src/Samplers/Sampler.cs
@@ -83,8 +83,8 @@ namespace RayTracer
             {
                 disk_samples.Add(new Point2D());
 
-                double r;
-                double phi;
+                float r;
+                float phi;
 
                 //Determine the quadrant for which to map according to the Shirley concentric map
                 //Quadrant 1 and 2:
@@ -94,13 +94,13 @@ namespace RayTracer
                     if(samples[i].x > samples[i].y)
                     {
                         r = samples[i].x;
-                        phi = (samples[i].y / samples[i].x) * (Math.PI / 4.0);
+                        phi = (samples[i].y / samples[i].x) * ((float)Math.PI / 4.0f);
                     }
                     //Otherwise in quadrant 2
                     else
                     {
                         r = samples[i].y;
-                        phi = (samples[i].y != 0.0) ? (2 - samples[i].x / samples[i].y) * (Math.PI / 4.0) : 0.0; //Don't divide by zero
+                        phi = (samples[i].y != 0.0f) ? (2 - samples[i].x / samples[i].y) * ((float)Math.PI / 4.0f) : 0.0f; //Don't divide by zero
                     }
                 }
                 //Quadrant 3 and 4
@@ -110,18 +110,18 @@ namespace RayTracer
                     if(samples[i].x < samples[i].y)
                     {
                         r = -samples[i].x;
-                        phi = (samples[i].x != 0.0) ? (4 + samples[i].y / samples[i].x) * (Math.PI / 4.0) : 0.0; //Don't divide by zero
+                        phi = (samples[i].x != 0.0f) ? (4 + samples[i].y / samples[i].x) * ((float)Math.PI / 4.0f) : 0.0f; //Don't divide by zero
                     }
                     //Otherwise in quadrant 4
                     else
                     {
                         r = -samples[i].y;
-                        phi = (samples[i].y != 0.0) ? (6 - samples[i].x / samples[i].y) * (Math.PI / 4.0) : 0.0; //Don't divide by zero
+                        phi = (samples[i].y != 0.0f) ? (6 - samples[i].x / samples[i].y) * ((float)Math.PI / 4.0f) : 0.0f; //Don't divide by zero
                     }
                 }
 
-                disk_samples[i].x = r * Math.Cos(phi);
-                disk_samples[i].y = r * Math.Sin(phi);
+                disk_samples[i].x = r * (float)Math.Cos(phi);
+                disk_samples[i].y = r * (float)Math.Sin(phi);
             }
         }
         public virtual void setup_shuffled_indices()
