+++ SCSRaytracer/src/Tracers/Whitted.cs
@@ -55,7 +55,7 @@ namespace RayTracer
                 }
             }
         }
-        public override RGBColor trace_ray(Ray ray, double tmin, int depth)
+        public override RGBColor trace_ray(Ray ray, float tmin, int depth)
         {
             throw new NotImplementedException();
         }
